etape 3 et 4 de epic 3

+++ src/main/java/com/wam/lab1_maintenance/model/Series.java
@@ -29,5 +29,5 @@ public class Series {
     private Integer nbEpisodes;
 
     @Column(name = "note")
-    private Double note;
+    private Float note;
 }
