+++ src/main/java/com/wam/lab1_maintenance/request/SeriesRequestBody.java
@@ -4,6 +4,6 @@ public record SeriesRequestBody(
     String title,
     String genre,
     Integer nbEpisodes,
-    Double note
+    Float note
 ) {}
